requirements.txt
discord.py==1.7.3
tomli==1.1.0
